Conditional clearing of finalizers
WithFinalizer's ReadyToClearFinalizer can be used to define custom rules
for clearing the finalizer. For example, deletion of a resource can be
blocked until all child resources are fully deleted replicating the
behavior of an owner reference in parent-child relationships that are
not supported by owner references. Client lookups and advanced logic
should be avoided as errors cannot be returned. Computed values can be
retrieved that were stashed from a previous reconciler, like a
SyncReconciler#Finalize hook.

Signed-off-by: Scott Andrews <scott@andrews.me>