ForEach reconciler
The ForEach reconciler can iterator over a slice calling a nested
reconciler for each item. The current item is stashed on the context to
be retrieved with a CursorStasher.

Nested iterators are allowed so long as the type of the items being
iterated over have unique names. We may enhance the generated stash key
in the future to relax this requirement, so users should not assume the
specific key to be stable between releases.

Signed-off-by: Scott Andrews <scott@andrews.me>