Allow rtesting to use custom diff func
Test cases can now specify a custom diff function which can be used to
either use custom cmp options, or replace cmp with some other differ.

The existing set of cmp options are matched using the DiffReason. Custom
funcs should assume additional DiffReasons will be defined in the
future. The DefaultDiff func is used if a custom Diff is not defined.
Custom funcs may delegate to the default for DiffReasons they choose not
to modify.

Signed-off-by: Scott Andrews <scott@andrews.me>