Quiet errors are always actually quiet
controller-runtime will always log an error returned it it from a
reconciler. In order to avoid the quieted errors from appearing in the
logs, we need to return a nil error. To get the same effect of an error
without the logging, we also request a requeue. The requeue is handled
the same as if the request erred in terms of the workqueue backoffs.

Signed-off-by: Scott Andrews <scott@andrews.me>