Suppress result when updating status (#609)
Updating the status for a resource will propagated from the api server
back to the informer causing the resource be added to the queue.
Directly requesting a requeue is counter productive because we may end
up reprocessing the now stale resource. Worse, requeues will extend the
work queue backoff while it should be reset.

We now always return an empty result after updating the status. The
resource is reprocessed organically.

ErrSkipStatusUpdate allows a sub reconciler to opt out of the status
update even if the value has changed. As an error, the request is
requeued automatically.

Signed-off-by: Scott Andrews <scott@andrews.me>