Allow opt-in to dangerous duck client operations (#506)
Duck types represent a partial view of a resource. It is dangerous to
perform operation that assume the client has the full resource when it
only has a portion of the resource. Calling update in this case can
result in data loss as fields that are not part of the duck type will be
dropped on the server. Likewise, it is difficult to create a resource
with a partial object as required/essential fields may not be available.

This change allows users to explicitly override these concerns and use
the UpdatingObjectManager to manage a child via a duck type. This should
only be done when the user is certain the duck typed object contains all
fields represented on the resource.

Signed-off-by: Scott Andrews <scott@andrews.me>