Support client.Apply() test expectations (#645)
Controller runtime added `Apply` as a client operation, which is sugar
around a patch operation. Testing with patch is a bit painful today as
the test needs to provide the json bytes of the patch. Diffs in the
expected and actual bytes are very annoying to debug. The Apply
operation uses a runtime.ApplyConfiguration object which can hold
semantic structure and should make failing tests easier to debug.

The ExpectConfig object now has an `ExpectApplies` slice that is
compared to the Apply operation recorded on the fake client for the
test. Likewise, an "apply" action is reacted during client requests. The
assertion compares the ApplyConfiguration object as a json map.

Signed-off-by: Scott Andrews <scott@andrews.me>