Silence conflicts on status update for ResourceReconciler
Conflict errors on status update are often benign and can only be
recovered by reprocessing the resource. As errors the can produce a
noisy log that distracts from more significant errors.

Because a high volume of status updates is indicative of a problem, an
info log line is captured instead of the full error. The
StatusUpdateFailed event is no longer recorded for conflicts.

Signed-off-by: Scott Andrews <scott@andrews.me>